refactor(benchmarks-website): split html.rs into html/ module
Split the 1080-line server/src/html.rs into a directory module:

- html/mod.rs        — router(), UiQuery, FilterState, parse_csv, the
  three async page handlers, and collect_landing_groups (entry
  surface).
- html/render.rs     — render_page, theme_bootstrap_script,
  site_header, the SVG icon helpers, error_page, versioned_asset and
  STATIC_ASSET_VERSION (page chrome).
- html/landing.rs    — landing_body, chart_card,
  downsample_badge_slot, LandingGroup.
- html/chart.rs      — chart_body, group_body (permalink page bodies).
- html/summary.rs    — summary_markup + format_time_ns.
- html/filter.rs     — filter_dropdown, filter_row, filter_state_script.
- html/toolbar.rs    — per_chart_toolbar, range_strip.
- html/static_assets.rs — serve_*, static_response, the
  include_bytes! consts, STATIC_ASSET_VERSION, versioned_asset.

Behaviour preserved: every snapshot test, integration assertion, and
inline JSON byte stays identical.
